backend2/requirements.txt
fastapi
uvicorn
duckdb
pandas
python-dotenv
groq
google-generativeai
plotly
